refactoring auth token support
