+++ pyproject.toml
@@ -34,6 +34,9 @@ allow-direct-references = true
 [tool.hatch.build.targets.wheel]
 packages = ["src/objectstore"]
 
+[project.scripts]
+authtoken = "objectstore.auth.tokens:main"
+
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "function"
